examples/kubeflow/mnist_with_tfjob/kubflowpach/.cache/manifests/manifests-1.0.0/tests/scripts/requirements.txt
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0